% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{Fycondx}
\alias{Fycondx}
\title{Fycondx}
\usage{
Fycondx(object, yvals, xdf)
}
\arguments{
\item{object}{either a distribution regression or quantile regression object}

\item{yvals}{the values to compute the ecdf for}

\item{xdf}{a dataframe (can contain multiple rows) with x values}
}
\value{
a list of conditional distributions
}
\description{
a generic method for computing conditional distributions
}
